+++ lib/inappstory_plugin.dart
@@ -3,7 +3,6 @@ import 'dart:async';
 import 'package:flutter/cupertino.dart';
 
 import 'inappstory_plugin_platform_interface.dart';
-import 'inappstory_sdk_module.dart';
 import 'src/controllers/feed_stories_controller.dart';
 import 'src/widgets/builders/builders.dart';
 import 'src/widgets/decorators/feed_decorator.dart';
@@ -13,11 +12,11 @@ import 'src/widgets/streams/feed_stories_stream.dart';
 export 'src/callbacks/callbacks.dart';
 export 'src/controllers/controllers.dart';
 export 'src/data/story.dart';
-export 'src/pigeon_generated.g.dart';
+export 'src/pigeon_generated.g.dart' hide InAppStoryManagerHostApi;
 export 'src/widgets/decorators/decorators.dart';
 export 'src/widgets/widgets.dart';
 
-class InAppStoryPlugin implements InAppStorySdkModule {
+class InAppStoryPlugin {
   factory InAppStoryPlugin() => _singleton ??= InAppStoryPlugin._private();
 
   InAppStoryPlugin._private();
@@ -25,9 +24,13 @@ class InAppStoryPlugin implements InAppStorySdkModule {
   static InAppStoryPlugin? _singleton;
 
   /// The [InAppStoryPlugin] initialization method.
-  @override
-  Future<void> initWith(String apiKey, String userId) {
-    return InappstoryPluginPlatform.instance.initWith(apiKey, userId);
+  Future<void> initWith(String apiKey, String userId, {Locale? locale}) async {
+    return InappstoryPluginPlatform.instance.initWith(
+      apiKey,
+      userId,
+      languageCode: locale?.languageCode,
+      languageRegion: locale?.countryCode,
+    );
   }
 
   @Deprecated('Use FeedStoriesWidget instead')