+++ lib/inappstory_sdk_module.dart
@@ -2,5 +2,6 @@ import 'dart:async';
 
 /// This is an abstract class that defines the interface for initializing the InAppStory SDK module.
 abstract class InAppStorySdkModule {
-  FutureOr<void> initWith(String apiKey, String userId);
+  FutureOr<void> initWith(String apiKey, String userId,
+      {String? languageCode, String? languageRegion});
 }