+++ lib/src/controllers/controllers.dart
@@ -1,2 +1,3 @@
 export 'feed_stories_controller.dart';
+export 'ias_manager.dart';
 export 'ias_statistics_manager.dart';